requirements.txt
Flask==3.1.2
Flask-Sock==0.7.0
google-cloud-speech==2.21.0
google-cloud-firestore==2.23.0
google-auth==2.47.0
grpcio==1.76.0
protobuf==6.33.4
python-dotenv==1.2.1
gunicorn==23.0.0
simple-websocket==1.1.0
